+++ android/src/test/java/com/revenuecat/purchases/hybridcommon/SkuDetailsMapperTests.kt
@@ -1,9 +1,7 @@
-package com.revenuecat.purchases.common
+package com.revenuecat.purchases.hybridcommon
 
-import com.revenuecat.purchases.models.ProductDetails
+import com.android.billingclient.api.SkuDetails
 import com.revenuecat.purchases.hybridcommon.mappers.mapIntroPrice
-import com.revenuecat.purchases.hybridcommon.mockCurrencyFormatter
-import com.revenuecat.purchases.hybridcommon.mockLogError
 import io.mockk.every
 import io.mockk.mockk
 import org.assertj.core.api.Assertions.assertThat
@@ -15,11 +13,11 @@ import org.junit.jupiter.api.Test
 internal class SkuDetailsMapperTests {
 
     var received: Map<String, Any?> = emptyMap()
-    val mockProductDetails = mockk<ProductDetails>(relaxed = true)
+    val mockSkuDetails = mockk<SkuDetails>(relaxed = true)
 
     @BeforeEach
     fun setup() {
-        every { mockProductDetails.priceCurrencyCode } returns "USD"
+        every { mockSkuDetails.priceCurrencyCode } returns "USD"
     }
 
     @Nested
@@ -28,8 +26,8 @@ internal class SkuDetailsMapperTests {
         @Test
         fun `of 7 days, the map has the correct intro price values`() {
             mockCurrencyFormatter(0, "$0.00")
-            every { mockProductDetails.freeTrialPeriod } returns "P7D"
-            received = mockProductDetails.mapIntroPrice()
+            every { mockSkuDetails.freeTrialPeriod } returns "P7D"
+            received = mockSkuDetails.mapIntroPrice()
             val expected = mapOf(
                 "price" to 0,
                 "priceString" to "$0.00",
@@ -44,8 +42,8 @@ internal class SkuDetailsMapperTests {
         @Test
         fun `of 1 month, the map has the correct intro price values`() {
             mockCurrencyFormatter(0, "$0.00")
-            every { mockProductDetails.freeTrialPeriod } returns "P1M"
-            received = mockProductDetails.mapIntroPrice()
+            every { mockSkuDetails.freeTrialPeriod } returns "P1M"
+            received = mockSkuDetails.mapIntroPrice()
             val expected = mapOf(
                 "price" to 0,
                 "priceString" to "$0.00",
@@ -60,8 +58,8 @@ internal class SkuDetailsMapperTests {
         @Test
         fun `of 0 days, the map has the correct intro price values`() {
             mockCurrencyFormatter(0, "$0.00")
-            every { mockProductDetails.freeTrialPeriod } returns "P0D"
-            received = mockProductDetails.mapIntroPrice()
+            every { mockSkuDetails.freeTrialPeriod } returns "P0D"
+            received = mockSkuDetails.mapIntroPrice()
 
             val expected = mapOf(
                 "price" to 0,
@@ -77,8 +75,8 @@ internal class SkuDetailsMapperTests {
         @Test
         fun `with value 365, the map has the correct intro price values`() {
             mockLogError()
-            every { mockProductDetails.freeTrialPeriod } returns "365"
-            received = mockProductDetails.mapIntroPrice()
+            every { mockSkuDetails.freeTrialPeriod } returns "365"
+            received = mockSkuDetails.mapIntroPrice()
             val expected = mapOf(
                 "price" to null,
                 "priceString" to null,
@@ -97,10 +95,10 @@ internal class SkuDetailsMapperTests {
     inner class MappingIntroPrice {
         @BeforeEach
         fun beforeEachTest() {
-            every { mockProductDetails.freeTrialPeriod } returns null
-            every { mockProductDetails.introductoryPriceAmountMicros } returns 10000000
-            every { mockProductDetails.introductoryPrice } returns "$10.00"
-            every { mockProductDetails.introductoryPriceCycles } returns 2
+            every { mockSkuDetails.introductoryPriceAmountMicros } returns 10000000
+            every { mockSkuDetails.introductoryPrice } returns "$10.00"
+            every { mockSkuDetails.introductoryPriceCycles } returns 2
+            received = mockSkuDetails.mapIntroPrice()
         }
 
         private val expectedCommon = mapOf(
@@ -111,8 +109,8 @@ internal class SkuDetailsMapperTests {
 
         @Test
         fun `of 7 days, the map has the correct intro price values`() {
-            every { mockProductDetails.introductoryPricePeriod } returns "P7D"
-            received = mockProductDetails.mapIntroPrice()
+            every { mockSkuDetails.introductoryPricePeriod } returns "P7D"
+            received = mockSkuDetails.mapIntroPrice()
             val expected = mapOf(
                 "period" to "P7D",
                 "periodUnit" to "DAY",
@@ -123,8 +121,8 @@ internal class SkuDetailsMapperTests {
 
         @Test
         fun `of 1 month, the map has the correct intro price values`() {
-            every { mockProductDetails.introductoryPricePeriod } returns "P1M"
-            received = mockProductDetails.mapIntroPrice()
+            every { mockSkuDetails.introductoryPricePeriod } returns "P1M"
+            received = mockSkuDetails.mapIntroPrice()
 
             val expected = mapOf(
                 "period" to "P1M",
@@ -136,8 +134,8 @@ internal class SkuDetailsMapperTests {
 
         @Test
         fun `of 0 days, the map has the correct intro price values`() {
-            every { mockProductDetails.introductoryPricePeriod } returns "P0D"
-            received = mockProductDetails.mapIntroPrice()
+            every { mockSkuDetails.introductoryPricePeriod } returns "P0D"
+            received = mockSkuDetails.mapIntroPrice()
 
             val expected = mapOf(
                 "period" to "P0D",
@@ -150,8 +148,8 @@ internal class SkuDetailsMapperTests {
         @Test
         fun `with a value of 365, the map has the correct intro price values`() {
             mockLogError()
-            every { mockProductDetails.introductoryPricePeriod } returns "365"
-            received = mockProductDetails.mapIntroPrice()
+            every { mockSkuDetails.introductoryPricePeriod } returns "365"
+            received = mockSkuDetails.mapIntroPrice()
 
             val expected = mapOf(
                 "price" to null,
@@ -167,9 +165,9 @@ internal class SkuDetailsMapperTests {
 
     @Test
     fun `"when mapping a SkuDetails with no free trial nor introductory price, the map has null intro price values`() {
-        every { mockProductDetails.freeTrialPeriod } returns ""
-        every { mockProductDetails.introductoryPrice } returns ""
-        received = mockProductDetails.mapIntroPrice()
+        every { mockSkuDetails.freeTrialPeriod } returns ""
+        every { mockSkuDetails.introductoryPrice } returns ""
+        received = mockSkuDetails.mapIntroPrice()
 
         val expected = mapOf(
             "price" to null,
